+++ tools/tpm2_pcrreset.c
@@ -7,7 +7,9 @@
 #include "tpm2.h"
 #include "tpm2_tool.h"
 #include "tpm2_options.h"
+#include "files.h"
 
+#define MAX_SESSIONS 3
 typedef struct tpm_pcr_reset_ctx tpm_pcr_reset_ctx;
 struct tpm_pcr_reset_ctx {
     /*
@@ -18,9 +20,19 @@ struct tpm_pcr_reset_ctx {
     /*
      * Outputs
      */
+
+    /*
+     * Parameter hashes
+     */
+    const char *cp_hash_path;
+    TPM2B_DIGEST cp_hash;
+    bool is_command_dispatch;
+    TPMI_ALG_HASH parameter_hash_algorithm;
 };
 
-static tpm_pcr_reset_ctx ctx;
+static tpm_pcr_reset_ctx ctx = {
+    .parameter_hash_algorithm = TPM2_ALG_ERROR,
+};
 
 static tool_rc pcr_reset(ESYS_CONTEXT *ectx) {
 
@@ -30,7 +42,8 @@ static tool_rc pcr_reset(ESYS_CONTEXT *ectx) {
             continue;
         }
 
-        tool_rc rc = tpm2_pcr_reset(ectx, i);
+        tool_rc rc = tpm2_pcr_reset(ectx, i, &ctx.cp_hash,
+            ctx.parameter_hash_algorithm);
         if (rc != tool_rc_success) {
             LOG_ERR("Could not reset PCR index: %ld", i);
             return rc;
@@ -47,6 +60,19 @@ static tool_rc process_outputs(ESYS_CONTEXT *ectx) {
     /*
      * 1. Outputs that do not require TPM2_CC_<command> dispatch
      */
+    bool is_file_op_success = true;
+    if (ctx.cp_hash_path) {
+        is_file_op_success = files_save_digest(&ctx.cp_hash, ctx.cp_hash_path);
+
+        if (!is_file_op_success) {
+            return tool_rc_general_error;
+        }
+    }
+
+    tool_rc rc = tool_rc_success;
+    if (!ctx.is_command_dispatch) {
+        return rc;
+    }
 
     /*
      * 2. Outputs generated after TPM2_CC_<command> dispatch
@@ -85,10 +111,21 @@ static tool_rc process_inputs(ESYS_CONTEXT *ectx) {
     /*
      * 4.a Determine pHash length and alg
      */
+    tpm2_session *all_sessions[MAX_SESSIONS] = {
+        0,
+        0,
+        0
+    };
+
+    const char **cphash_path = ctx.cp_hash_path ? &ctx.cp_hash_path : 0;
+
+    ctx.parameter_hash_algorithm = tpm2_util_calculate_phash_algorithm(ectx,
+        cphash_path, &ctx.cp_hash, 0, 0, all_sessions);
 
     /*
      * 4.b Determine if TPM2_CC_<command> is to be dispatched
      */
+    ctx.is_command_dispatch = ctx.cp_hash_path ? false : true;
 
     return tool_rc_success;
 }
@@ -121,9 +158,26 @@ static bool on_arg(int argc, char** argv) {
     return true;
 }
 
+static bool on_option(char key, char *value) {
+
+    switch (key) {
+    case 0:
+        ctx.cp_hash_path = value;
+        break;
+    }
+
+    return true;
+}
+
 static bool tpm2_tool_onstart(tpm2_options **opts) {
 
-    *opts = tpm2_options_new(NULL, 0, NULL, NULL, on_arg, 0);
+    static const struct option topts[] = {
+        { "cphash", required_argument, 0, 0},
+
+    };
+    
+    *opts = tpm2_options_new(NULL, ARRAY_LEN(topts), topts, on_option, on_arg,
+        0);
     return *opts != NULL;
 }